renamed cloud-energy-collection dir to iriscasttols
changed file structure to fix pylint error
